Settings: Change 'Model Number' to 'Model Name'
Most devices have the Model Number set to something like 'One' rather
than 'PN072', therefore they are using the Model Name not the Model
Number. Therefore, either every device should update to properly refer
to the Model Number, or we should update Settings to correctly state
that it is the Model Name.

Change-Id: I70981d3a46e7a7e26f4bfcb6c5d768a00314e3b9

Conflicts:
	res/values/cm_strings.xml